Implement Daemon test context manager
Address #39

Implement Daemon test context manager to automatically shutdown
running watchers when the tests fail.
