+++ INSTALLATION.md
@@ -1,6 +1,14 @@
 # Installing CoRR-cli
 
-To install CoRR-cmd, clone the repository and install:
+## Requirements
+
+The requirements are outlined in [requirements.txt](requirements.txt).
+Currently CoRR-cli only works on linux due to the `deamon` package
+only working on linux.
+
+## Installation
+
+To install CoRR-cli, clone the repository and install:
 
     $ git clone git@github.com:usnistgov/corr-cli.git
     $ cd corr-cli
