+++ corrcli/tests/test_task_manager.py
@@ -9,7 +9,7 @@ from click.testing import CliRunner
 from corrcli.corr_daemon import CoRRDaemon
 from corrcli.commands.cli import DEFAULT_WRITE_REFRESH_RATE
 from corrcli.stores.file_store import FileStore
-from corrcli.tools import start_daemon
+from corrcli.corr_daemon import start_daemon
 
 
 def configure(config_dir, refresh_rate):
@@ -23,19 +23,8 @@ def configure(config_dir, refresh_rate):
                  '--email=test@test.com',
                  '--author="Test Person"',
                  '--write-refresh-rate={0}'.format(refresh_rate)]
-    Popen(arguments).wait()
+    Popen(arguments).communicate()
     sleep(3)
-
-def stop_daemon(config_dir):
-    """Stop all daemons
-
-    """
-    arguments = ['corrcli',
-                 '--config-dir={0}'.format(config_dir),
-                 'watch',
-                 'stop',
-                 '--all']
-    Popen(arguments)
 
 def start_process(daemon_id, config_dir):
     """Launch a test process.
@@ -62,24 +51,21 @@ def test_task_manager():
     runner = CliRunner()
     with runner.isolated_filesystem() as config_path:
         configure(config_path, write_refresh_rate)
-        start_daemon(config_path)
-        daemon_df = CoRRDaemon.list(config_path)
-        daemon_id = daemon_df.daemon_id[0]
-        try:
-            process = start_process(daemon_id, config_path)
+        with start_daemon(config_path) as daemon_id:
+            try:
+                process = start_process(daemon_id, config_path)
+                sleep(write_refresh_rate + 1)
+                tasks = FileStore.read_all(config_path)
+            except: # pragma: no cover
+                process.kill()
+                raise
+            process.kill()
+            assert len(tasks) == 1
+            assert tasks[0]['status'] != 'finished'
             sleep(write_refresh_rate + 1)
             tasks = FileStore.read_all(config_path)
-        except: # pragma: no cover
-            process.kill()
-            raise
-        process.kill()
-        assert len(tasks) == 1
-        assert tasks[0]['status'] != 'finished'
-        sleep(write_refresh_rate + 1)
-        tasks = FileStore.read_all(config_path)
-        assert len(tasks) == 1
-        assert tasks[0]['status'] == 'zombie'
-        stop_daemon(config_path)
+            assert len(tasks) == 1
+            assert tasks[0]['status'] == 'zombie'
 
 
 def test_noconfig():
@@ -87,17 +73,14 @@ def test_noconfig():
     """
     runner = CliRunner()
     with runner.isolated_filesystem() as config_path:
-        start_daemon(config_path)
-        daemon_df = CoRRDaemon.list(config_path)
-        daemon_id = daemon_df.daemon_id[0]
-        try:
-            process = start_process(daemon_id, config_path)
-            sleep(DEFAULT_WRITE_REFRESH_RATE * 2)
-            tasks = FileStore.read_all(config_path)
-        except: # pragma: no cover
+        with start_daemon(config_path) as daemon_id:
+            try:
+                process = start_process(daemon_id, config_path)
+                sleep(DEFAULT_WRITE_REFRESH_RATE * 2)
+                tasks = FileStore.read_all(config_path)
+            except: # pragma: no cover
+                process.kill()
+                raise
             process.kill()
-            raise
-        process.kill()
-        assert len(tasks) == 1
-        assert tasks[0]['status'] != 'finished'
-        stop_daemon(config_path)
+            assert len(tasks) == 1
+            assert tasks[0]['status'] != 'finished'
