+++ corrcli/tools/__init__.py
@@ -3,7 +3,6 @@
 from subprocess import Popen
 from time import sleep
 import os
-
 import click
 
 
@@ -37,20 +36,3 @@ def get_config_dir(app_name): # pragma: no cover
     if not os.path.exists(config_dir):
         os.makedirs(config_dir)
     return config_dir
-
-def start_daemon(config_dir, callback_func=None):
-    """Start a daemon from a separate process.
-
-    Args:
-      config_dir: the CoRR configuration directory
-      callback_func: the name of the callback function
-    """
-    arguments = ['corrcli',
-                 '--config-dir={0}'.format(config_dir),
-                 'watch',
-                 'start',
-                 '--log']
-    if callback_func is not None:
-        arguments.append('--callback-func={0}'.format(callback_func))
-    Popen(arguments)
-    sleep(3)
